+++ lib/bugsnag/report.rb
@@ -23,11 +23,13 @@ module Bugsnag
     attr_accessor :api_key
     attr_accessor :app_type
     attr_accessor :app_version
+    attr_accessor :breadcrumbs
     attr_accessor :configuration
     attr_accessor :context
     attr_accessor :delivery_method
     attr_accessor :exceptions
     attr_accessor :hostname
+    attr_accessor :runtime_versions
     attr_accessor :grouping_hash
     attr_accessor :meta_data
     attr_accessor :raw_exceptions
@@ -51,8 +53,10 @@ module Bugsnag
       self.api_key = configuration.api_key
       self.app_type = configuration.app_type
       self.app_version = configuration.app_version
+      self.breadcrumbs = []
       self.delivery_method = configuration.delivery_method
       self.hostname = configuration.hostname
+      self.runtime_versions = configuration.runtime_versions.dup
       self.meta_data = {}
       self.release_stage = configuration.release_stage
       self.severity = auto_notify ? "error" : "warning"
@@ -95,7 +99,8 @@ module Bugsnag
         },
         context: context,
         device: {
-          hostname: hostname
+          hostname: hostname,
+          runtimeVersions: runtime_versions
         },
         exceptions: exceptions,
         groupingHash: grouping_hash,
@@ -110,7 +115,14 @@ module Bugsnag
       payload_event = Bugsnag::Cleaner.clean_object_encoding(payload_event)
 
       # filter out sensitive values in (and cleanup encodings) metaData
-      payload_event[:metaData] = Bugsnag::Cleaner.new(configuration.meta_data_filters).clean_object(meta_data)
+      filter_cleaner = Bugsnag::Cleaner.new(configuration.meta_data_filters)
+      payload_event[:metaData] = filter_cleaner.clean_object(meta_data)
+      payload_event[:breadcrumbs] = breadcrumbs.map do |breadcrumb|
+        breadcrumb_hash = breadcrumb.to_h
+        breadcrumb_hash[:metaData] = filter_cleaner.clean_object(breadcrumb_hash[:metaData])
+        breadcrumb_hash
+      end
+
       payload_event.reject! {|k,v| v.nil? }
 
       # return the payload hash
@@ -151,6 +163,26 @@ module Bugsnag
     # Tells the client this report should not be sent.
     def ignore!
       @should_ignore = true
+    end
+
+    ##
+    # Generates a summary to be attached as a breadcrumb
+    #
+    # @return [Hash] a Hash containing the report's error class, error message, and severity
+    def summary
+      # Guard against the exceptions array being removed/changed or emptied here
+      if exceptions.respond_to?(:first) && exceptions.first
+        {
+          :error_class => exceptions.first[:errorClass],
+          :message => exceptions.first[:message],
+          :severity => severity
+        }
+      else
+        {
+          :error_class => "Unknown",
+          :severity => severity
+        }
+      end
     end
 
     private